recipe-rag/recipe-rag/backend/config.py
# Environment configuration loading
import os
from dotenv import load_dotenv
import logging

# Load environment variables
load_dotenv()

logger = logging.getLogger(__name__)

class Config:
    # OpenAI Configuration
    OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
    
    # Chroma Configuration
    CHROMA_PERSIST_DIRECTORY = os.getenv("CHROMA_PERSIST_DIRECTORY", "./vectordata")
    
    # Logging Configuration
    LOG_LEVEL = os.getenv("LOG_LEVEL", "INFO")
    LOG_FILE = os.getenv("LOG_FILE", "logs/backend.log")
    
    # Recipe Configuration
    RECIPES_DIRECTORY = os.getenv("RECIPES_DIRECTORY", "data/recipes")
    
    @classmethod
    def validate(cls):
        """Validate that required configuration is present"""
        if not cls.OPENAI_API_KEY:
            logger.warning("OPENAI_API_KEY not set in environment")
        
        logger.info("Configuration loaded successfully")